pull dialer into net.dialer
